requirements/base.txt
# Django
django==3.2
djangorestframework==3.13
django-environ==0.9.0
djangorestframework-simplejwt==5.2.2

requests==2.28.2
psycopg2==2.9.5

django-storages==1.12.3
cachetools==4.2.4
google-api-core==2.7.2
google-auth==2.6.5
google-cloud-core==2.3.0
google-cloud-storage==2.0.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
django-cleanup==7.0.0